23477 Fix aeat sii report
At tryton5.2 checked if the user was root to return every record from the pool, otherwise it applied the domain to fetch the results.

Since the issue4080, now checks if the if the user is root (this continues) or if the transaction context _check_access is None or False.

This is not an issue from the views (for example), because they make RPC calls, which set the _check_access=True.

But from a wizard there are no RPC calls, and the transaction context doesn't have check_access attribute, hence the domain becomes all the records (which creates multiple reports.